+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0 BETA 11] - 2019-04-04
+### Fixed
+- 4-byte characters such as emojis would cause a database error. To fix I change them to a single 3-byte character.
+
+## [1.0.0 BETA 10] - 2019-01-08
+### Added
+- Czech translation (email template files only)
+
 ## [1.0.0 BETA 9] - 2018-08-17
 ### Changed
 - Replaced my custom email template manager with the built-in one